# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AR Studio API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/variants:
    get:
      operationId: listVariants
      tags: [variants]
      summary: List variants for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of variants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectVariant"

    post:
      operationId: createVariant
      tags: [variants]
      summary: Create a variant for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVariantBody"
      responses:
        "201":
          description: Variant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectVariant"
  /projects/{projectId}/variants/{id}:
    patch:
      operationId: updateVariant
      tags: [variants]
      summary: Update a variant
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVariantBody"
      responses:
        "200":
          description: Updated variant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectVariant"
        "404":
          description: Not found

    delete:
      operationId: deleteVariant
      tags: [variants]
      summary: Delete a variant
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    ProjectVariant:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        modelUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - thumbnailUrl
        - modelUrl
        - sortOrder
        - createdAt
    CreateVariantBody:
      type: object
      properties:
        name:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        modelUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
      required:
        - name
    UpdateVariantBody:
      type: object
      properties:
        name:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        modelUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer